requirements.txt
django
djangorestframework
rembg
onnxruntime
Pillow
requests